requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.1.11
Django==4.0
django-allauth==0.48.0
django-cors-headers==3.10.0
django-environ==0.9.0
djangorestframework==3.13.0
drf-spectacular==0.21.0
environs==9.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==5.3.0
